Following changes on the classifier tools for day1 exercises
The new code runs on the day1.py script. The only exception is that I could not run

classifiers/max_ent_online.py

due to incompatibilities with the scipy version I am currently running. This should be solved soon.

Added lots of comments and extra spacing so that newcomers can quickly uncerstand what the code does.
Homogenized terms and code across differen methods e.g. "nr_epochs" replaced by "round_nr" used at first models. Also parameter initialization and similar common stages in the algorithms were given the similar structure and names.

In particular in

classifiers/gaussian_naive_bayes.py

I removed the use of estimate_gaussian, since it is just a mean calculation. I think this way it is easyer to follow the guide.
I also removed the variance which was not used.

in

classifiers/multinomial_naive_bayes.py

Replaced the variables likelihoods and sums by class_count and total_count which where more coherent whith what it is explained in the guide

IMPORTANT

I changed the computation of the learning rate for maxent and svn. It is now computed outside of the feature loop and it does not decrease with each feature loop iteration. This is my interpretation of the indications on the guide but I could be wrong.