Partially solved Warning/nan in Exercise 2.9
These are caused by the log-domai operations, which lead to not-a-numbers (nans) in posterior decoding. The recognition rates reported are thus not correct. This has been solved by changing the logzero() to -1e20 and using safe_log() to treat the case log(0) separately.

The performance of posterior decoding is now much higher. I get

Test Set Accuracy: Posterior Decode 0.862, Viterbi Decode: 0.867