Clarification of 'reward' variable meaning in RL lab
It was said in guide in section 6.4 that rewards [10., 2., 3.] are rewards for a transition into state 1, 2, and 3, respectively.
But such definition of 'rewards' variable is not consistent with code of solution, despite code of solution is correct.
So, here we change definition of 'rewards' variable to make it consistent with code.
New definition is:
'rewards' variable contains expected values of the next reward for each state
And naturally, for clarity we also introduce new variable - 'raw_rewards':
'raw_rewards' variable contains rewards obtained after transition to each state.
As for 'rewards' variable, now we generate it from 'raw_rewards' and 'policy', instead of assigning directly.
For more details see corresponding commit in 'lxmls-guide' project.